+++ code/cell.ui.sys/src/components.dev/ModuleView.dev/module.Sample/Module.tsx
@@ -1,7 +1,7 @@
 import { filter } from 'rxjs/operators';
 
 import { Module, t } from '../common';
-import { renderer } from './view/render';
+import { renderer } from './Module.render';
 
 type P = t.MyProps;
