+++ code/cell.ui.sys/src/components.dev/ModuleView.dev/module.Sample/components/404.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { color, COLORS, css, CssValue } from '../../../common';
+import { color, COLORS, css, CssValue } from '../../common';
 import { Identifiers } from './Identifiers';
 
 export type ITest404Props = {
