+++ code/cell.ui.sys/src/components.dev/ModuleView.dev/module.Sample/components/Diagram.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
-import { COLORS, css, CssValue, t } from '../../../common';
+import { COLORS, css, CssValue, t } from '../../common';
 
 export type ITestDiagramProps = {
   module?: t.IModule;
