+++ code/cell.ui.sys/src/components.dev/ModuleView.dev/module.Sample/components/Identifiers.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { COLORS, css, CssValue } from '../../../common';
+import { COLORS, css, CssValue } from '../../common';
 
 export type IIdentifiersProps = {
   module?: string;
