+++ code/cell.ui.sys/src/components.dev/ModuleView.dev/module.Sample/components/TreeColumns.tsx
@@ -5,7 +5,7 @@ import { debounceTime, takeUntil } from 'rxjs/operators';
 import { Button } from '@platform/ui.button';
 
 import { Tree } from '@platform/ui.tree';
-import { t } from '../../../common';
+import { t } from '../../common';
 
 type Node = t.ITreeviewNode;
